model_training/requirements.txt
# ── ML / training ─────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
pillow>=9.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
optuna>=3.0

# ── Notebook environment ──────────────────────────────────────────────────
jupyter>=1.1.0
